LayoutTests/platform/ios-sim-deprecated/http/tests/security/cross-origin-shared-worker-expected.txt

This iframe should return a security error:

This iframe should not return any errors:



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
SecurityError

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Connected successfully
